add color, cleanup, misc
