+++ programs/castle-vault/src/lib.rs
@@ -1,6 +1,6 @@
 use anchor_lang::prelude::*;
 
-use crate::borsh::{BorshDeserialize, BorshSerialize};
+// use crate::borsh::{BorshDeserialize, BorshSerialize};
 
 pub mod adapters;
 pub mod backend_container;
